requirements.txt
python-socketio
gevent
gevent-websocket
ujson
